fix(wiki): skip H1 in SectionedMarkdownBody to stop double-render
parseSectionsFromMarkdown follows the standard "next same-or-higher
heading" rule, so an H1 with no following H1 gets endLine = EOF. The
render loop in SectionedMarkdownBody previously rendered level-1 via
the full-span else branch (startLine..endLine+1 inclusive), which
blasted the entire document as one block. Then the loop continued to
each H2+, rendering body-only — every post-H1 section rendered twice.

The wiki title is already rendered as page chrome by WikiEntityArticle,
so the markdown-level H1 is redundant. Skip level-1 in the loop and
the preamble handles any pre-H1 text on the rare non-first-line case.

Extract SectionedMarkdownBody (plus SectionHeadingWithEdit and helpers)
from page.tsx into its own module so it can be rendered in a test
without booting the whole detail page. New render-time regression test
asserts post-H1 content appears exactly once for both the H1-leading
and no-H1-preamble paths.

Fixes #152.
