+++ wiki/src/app/(shell)/wiki/[id]/page.tsx
@@ -24,6 +24,7 @@ import { WikiInfobox } from "@/components/wiki/WikiInfobox";
 import { WikiChip } from "@/components/wiki/WikiChip";
 import { WikiCitations } from "@/components/wiki/WikiCitations";
 import { WikiEditLink } from "@/components/wiki/WikiFurniture";
+import { SectionedMarkdownBody } from "./SectionedMarkdownBody";
 import {
   parseSectionsFromMarkdown,
   replaceSectionInMarkdown,
@@ -113,191 +114,6 @@ function HtmlWikiBody({
       dangerouslySetInnerHTML={{ __html: html }}
     />
   );
-}
-
-/**
- * Match server-computed `sections[]` entries to the anchors produced by
- * `parseSectionsFromMarkdown` against the displayed markdown. The server
- * slugifier and the client helper are kept in sync (see
- * `wiki/src/lib/sectionEdit.ts` header) so this lookup is a plain
- * `Map<anchor, WikiSection>`.
- */
-function buildCitationsByAnchor(
-  sections: WikiSection[] | undefined,
-): Map<string, WikiSection> {
-  const map = new Map<string, WikiSection>();
-  if (!sections) return map;
-  for (const s of sections) {
-    map.set(s.anchor, s);
-  }
-  return map;
-}
-
-/**
- * Heading styles mirroring `MarkdownContent`'s internal `buildComponents`
- * heading mapping so that section-scoped headings (rendered by this
- * page) look identical to headings rendered inside `<MarkdownContent>`
- * body blocks. Kept inline rather than extracted to a shared module so
- * the duplication is local and easy to resync when either side drifts.
- */
-const sectionHeadingStyle: Record<2 | 3 | 4, CSSProperties> = {
-  2: {
-    ...T.h2,
-    color: "var(--wiki-article-h2)",
-    margin: "24px 0 8px",
-    borderBottom: "1px solid var(--wiki-card-border)",
-    paddingBottom: 4,
-  },
-  3: {
-    ...T.h3,
-    color: "var(--wiki-article-h2)",
-    margin: "20px 0 6px",
-  },
-  4: {
-    ...T.h4,
-    color: "var(--wiki-article-h2)",
-    margin: "16px 0 4px",
-  },
-};
-
-/**
- * Render a section's heading line with a trailing `[edit]` affordance.
- * The edit link is suppressed for H1 (per the phase spec — editing H1
- * is equivalent to full-body edit, which lives on the Edit tab) and for
- * H5/H6 (no visual treatment exists in `MarkdownContent`; fall through
- * to plain `MarkdownContent` rendering).
- *
- * Returns `null` when the level is outside 2–4 so the caller falls back
- * to rendering the whole section — including its heading — via
- * `<MarkdownContent>`.
- */
-function SectionHeadingWithEdit({
-  section,
-  onEdit,
-  showEditLink,
-}: {
-  section: SectionInfo;
-  onEdit: (sectionId: string) => void;
-  showEditLink: boolean;
-}) {
-  const style = sectionHeadingStyle[section.level as 2 | 3 | 4];
-  if (!style) return null;
-
-  const HeadingTag = (section.level === 3
-    ? "h3"
-    : section.level === 4
-      ? "h4"
-      : "h2") as "h2" | "h3" | "h4";
-
-  const editLink = showEditLink ? (
-    <>
-      {" "}
-      <WikiEditLink onClick={() => onEdit(section.id)} />
-    </>
-  ) : null;
-
-  return (
-    <HeadingTag style={style}>
-      {section.heading}
-      {editLink}
-    </HeadingTag>
-  );
-}
-
-/**
- * Render the markdown body as a sequence of section-scoped
- * `<MarkdownContent>` blocks, each followed by its `<WikiCitations>`
- * superscripts. Preamble before the first heading (if any) renders as
- * an unattributed leading block.
- *
- * If the body has no headings, falls back to a single whole-body render
- * — `sections` is empty in that case, so no citations are rendered.
- *
- * When `onEditSection` is provided, H2/H3/H4 headings gain a trailing
- * `[edit]` bracket affordance. H1 is excluded (editing it is equivalent
- * to whole-body edit). The heading itself is extracted and rendered
- * separately so the `[edit]` link can sit next to the heading text;
- * only the section body (lines after the heading) goes through
- * `<MarkdownContent>`.
- */
-function SectionedMarkdownBody({
-  content,
-  refs,
-  sections,
-  style,
-  onEditSection,
-}: {
-  content: string;
-  refs: Record<string, WikiRef>;
-  sections: WikiSection[] | undefined;
-  style: CSSProperties;
-  onEditSection?: (sectionId: string) => void;
-}) {
-  const parsed: SectionInfo[] = parseSectionsFromMarkdown(content);
-  if (parsed.length === 0) {
-    return <MarkdownContent content={content} refs={refs} style={style} />;
-  }
-
-  const lines = content.split("\n");
-  const citationsByAnchor = buildCitationsByAnchor(sections);
-
-  const preamble = lines.slice(0, parsed[0].startLine).join("\n");
-  const blocks: ReactNode[] = [];
-  if (preamble.trim().length > 0) {
-    blocks.push(
-      <MarkdownContent
-        key="__preamble"
-        content={preamble}
-        refs={refs}
-        style={style}
-      />,
-    );
-  }
-
-  for (const section of parsed) {
-    const matched = citationsByAnchor.get(section.anchor);
-    const citations = matched?.citations ?? [];
-
-    // H1 never gets the [edit] affordance — it's the document-level
-    // heading and section-editing it is equivalent to full-body edit.
-    // Levels outside 2–4 fall through to plain MarkdownContent so the
-    // `[edit]` affordance isn't shown on H5/H6 that MarkdownContent
-    // renders with default styling.
-    const canExtractHeading =
-      section.level >= 2 && section.level <= 4 && onEditSection !== undefined;
-
-    if (canExtractHeading) {
-      const bodyOnly = lines
-        .slice(section.startLine + 1, section.endLine + 1)
-        .join("\n");
-      blocks.push(
-        <div key={section.anchor} id={section.anchor}>
-          <SectionHeadingWithEdit
-            section={section}
-            onEdit={onEditSection}
-            showEditLink={true}
-          />
-          {bodyOnly.trim().length > 0 && (
-            <MarkdownContent content={bodyOnly} refs={refs} style={style} />
-          )}
-          {citations.length > 0 && <WikiCitations citations={citations} />}
-        </div>,
-      );
-      continue;
-    }
-
-    const body = lines
-      .slice(section.startLine, section.endLine + 1)
-      .join("\n");
-    blocks.push(
-      <div key={section.anchor} id={section.anchor}>
-        <MarkdownContent content={body} refs={refs} style={style} />
-        {citations.length > 0 && <WikiCitations citations={citations} />}
-      </div>,
-    );
-  }
-
-  return <>{blocks}</>;
 }
 
 export default function WikiDetailPage() {
